+++ app/src/main/java/com/example/testtaskintegrio/domain/model/PointDomain.kt
@@ -8,5 +8,5 @@ data class PointDomain(
     val corpus: Long,
     val geoHash: String,
     val coordinates: GeoPoint,
-    val properties: Map<String, Int>
+    val properties: Map<String, Long>
 )
